Add JWT service and user registration (#28)
A JWT service was added to the `ServicesAndRepositoryExtension.cs` to handle JWT token generation. The `AuthController` was also extended with an email confirmation check to increase security. A user registration method was put in place, which validates the model, creates a new ApplicationUser, and checks for successful creation. This update increases security and allows for user registration.