Add cycle CRUD functionality and Map system
This commit has established full CRUD functionality for the "Cycle" entity in the system. The Dto models for Create, Update and viewing of Cycle objects have been added. A repository and a service for the 'Cycle' entity have been introduced. In addition, a Cycle controller was created with an endpoint for accessing all Cycle data, sorted and paginated as necessary. A mapping system has also been put into place using AutoMapper, allowing for the successful translation between Cycle entities and Cycle DTOs. 'CycleRepository' and 'CycleService' were registered in 'ServicesAndRepositoryExtension'. Npgsql was used for working with PostgreSQL for the EF Core database. This functionality was necessary to enable users to view, create, update and delete cycle data within the application. Changes are also included for the AutoMapper and AutoMapperExtensions to include mappings and mapping methods for CycleDto, CreateCycleDto, UpdateCycleDto and PaginatedList of CycleDto. Application now includes 'Paginator`, `PaginatedList` for handling paginated responses on 'Cycle' requests.