Refactor codebase and improve test suite
Refactored the directory structure and namespaces for Cycle-related DTOs and made a few changes in the Identity-related implementation for better consistency. A new file RegisterRequestDto.cs was created for better segregation of concerns. Similarly, the Cycle-related DTOs were moved to a new directory for better structured organization of these files.

Moreover, the test suite was extensively enhanced. Key changes include conversion of the test suite from NUnit to Xunit, utilization of a PostgresSQL container for integration testing, and the introduction of comprehensive tests for Cycle and Authentication functionalities.

These changes are aimed at improving the maintainability and readability of the codebase, enabling easier addition of new features in future and strengthening the robustness of the test suite.