Add AuthException and use it in AuthController
Added a new AuthException class in the Models/Exception/ directory and utilized it in our AuthController class. This update helps to provide clearer and more specific error messages to our users when authenticating, increasing the applications transparency and usability. Revised the HTTP status code in the integration tests, and updated the ExceptionFilter to handle AuthException. The goal is to return "Bad Request" instead of "Internal Server Error" when the user authentication fails, making the response more accurate.