Add forgot and reset password functionality
A new account controller is added with methods for resetting and forgetting passwords. Also provided corresponding Dto classes: ResetPasswordDto and ForgotPasswordDto.
Added the ResetPasswordException in the ExceptionFilter to handle password reset exceptions. Also updated the IEmailService interface and its implementation to include the SendForgotPasswordEmailAsync method.
Email exceptions are now handled in the ExceptionFilter by adding a new EmailVerifyException case.
All these changes were necessary to offer users the ability to reset their passwords and enhance the security of the application.