Refactor Cycle creation and CalendarDay state
In this commit, the process of creating a Cycle has been refactored by removing the need for a CreateCycleDto object as a parameter. Instead, a new cycle is being created with the current date and an unset end date.
State property in the CalendarDay model was  updated to "CycleDay" from "Period" for clear understanding.
Furthermore, a new function CreteCycleDay was introduced to the ICalendarDayRepository interface and implemented in the CalendarDayRepository class. This method creates a new cycle day by getting the last cycle of the user, throwing an exception if no previous cycle is found.
In the MetricService class, instead of throwing a NotFoundException when no calendar day is found, the new CreteCycleDay function is now being called.
This commit improves the process of creating new cycles and cycle days.