Add chart.js library and update analytics UI
Modified the file structure and components to incorporate chart.js library for graphing user cycle data in the analytics section. The UI was refactored for better readability and layout of information. The condition to display the flex div in card component was also tweaked to resolve an empty div issue.