Add custom encryption for user data protection
Added CustomLookupProtector and CustomLookupProtectorKeyRing classes for encryption and decryption of user data with symmetric key algorithm. Updated IdentityBuilder to use the new classes for protecting personal data. Additional keys and settings are added in appsettings files for different environments to initialize the key ring settings. Include these changes to enhance data security measures in the application.