Add production environment settings
A new environment file for production has been added and the angular.json file has been updated to replace environment.ts with environment.prod.ts when building for production. This change allows the app to use production-specific settings such as base URL, user email and password, when it is built in the production mode.