Upgrade packages and switch from NUnit to Xunit in UnitTests
The commit upgrades Microsoft, Dapper and Npgsql packages to their latest versions across the Models, Infrastructure, Vital, IntegrationTests and UnitTests subprojects. In addition, it replaces NUnit with Xunit in the UnitTests project. Changes in parsing DateTime to DateTimeOffset are also included in this commit. The upgrades keep the projects up-to-date while changing the testing framework prepares for simplified and consistent test writing in the future.